// src/accounts/account.enums.ts
export enum AccountType {
  SAVINGS = 'SAVINGS',
  CHECKING = 'CHECKING',
  BUSINESS_CURRENT = 'BUSINESS_CURRENT',
}

export enum AccountCurrency {
  USD = 'USD',
  FC = 'FC',
}

export enum AccountStatus {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  SUSPENDED = 'SUSPENDED',
  DORMANT = 'DORMANT',
  CLOSED = 'CLOSED',
}